requirements.txt
streamlit
pandas
selenium
webdriver-manager
line-bot-sdk